agent-standalone/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pyjwt==2.8.0
anthropic==0.34.0